pam/integration-tests: Move main integration test to generic file
