+++ pam/integration-tests/cli_test.go
@@ -2,9 +2,7 @@ package main_test
 
 import (
 	"context"
-	"flag"
 	"fmt"
-	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -254,24 +252,4 @@ func saveArtifactsForDebug(t *testing.T, artifacts []string) {
 			t.Logf("Could not write artifact %q: %v", artifact, err)
 		}
 	}
-}
-
-func TestMain(m *testing.M) {
-	// Needed to skip the test setup when running the gpasswd mock.
-	if os.Getenv("GO_WANT_HELPER_PROCESS") != "" {
-		os.Exit(m.Run())
-	}
-
-	testutils.InstallUpdateFlag()
-	flag.Parse()
-
-	execPath, daemonCleanup, err := testutils.BuildDaemon("-tags=withexamplebroker,integrationtests")
-	if err != nil {
-		log.Printf("Setup: Failed to build authd daemon: %v", err)
-		os.Exit(1)
-	}
-	defer daemonCleanup()
-	daemonPath = execPath
-
-	m.Run()
 }
